Create a dockerfile for angular cli
This Dockerfile is meant to be used for development purposes.
It generates a container that can run ng command for angular-cli.
It contains the necessary dependencies in order to work from everywhere and avoid polluting each different working computer.
It will probably need to be updated as the project goes on and more packages are added.