Change python base image and dependencies installation
The base image is slim-buster instead of buster.
Dependencies are installed in a single layer.
In order to install the aforementioned application dependencies some implicit dependencies are installed and immediately uninstalled in the same layer.
This way the created image size is kept small.
Base python server image size reduced from 946MB to 209MB.