First step of generalizing compile function
Function compile() is not bound to C or C++ compilation anymore.
It is more generalized so that it gets as arguments the upload path and the parser for compilation options json.
The api for C and C++ does not call immediately the specific compile function.
It rather calls the generic compile() function using the C/C++ specific arguments.