Login navigates to redirected_from url
If a non-logged-in user tries to access an authorised path, he is redirected to login page (due to guard).
The path that the user tried to access is stored as a parameter in the url.
If the user logs in successfully he is not redirected to the initial page but to this authorised path.