Add api route that implements both 'get results zip' and 'get personal file content'
The results zip and the source code file are (of course) just a file.
Thus getting a source code file and getting a zip file can be implemented from the same api with different query parameters.
The content type in response will be different (application/zip and text/plain respectively).
The frontend will have to differentiate the responses and handle them accordingly in each case.
Remove the existing api route for 'get results zip'.
Rename the api route for 'get all personal files'.