Frontend: Inject SourceCodeFileService in personal-files component
Hardcoded files dictionary is removed.
The variables are initialized in the constructor, so that they don't have undefined value.
The variable "files" get its value via subscription to sourceCodeFileService function getPersonalFiles() in ngOnInit.
Variable "files" is filled with any missing language and the other variables are filled using "files".
Variable "loading" exists as a flag for files download completion.