Backend: Module compile_c implements a subclass CCompilationHandler of CompilationHandler abstract class
The existing functions in compile_c module have become the implementation of the abstract methods.
There is no change in the code in these functions.
Of course, the names are changed in order to match the abstract methods.
The api routes for C compilation have not changed.
The implementation of the api routes for C compilation has changed.
They create a CCompilationHandler object and use its methods as arguments in compile() function.