Backend: Module compile_cpp implements a subclass CppCompilationHandler of CompilationHandler abstract class
The existing functions in compile_cpp module have become the implementation of the abstract methods.
There is no change in the code in these functions.
Of course, the names are changed in order to match the abstract methods.
The api routes for Cpp compilation have not changed.
The implementation of the api routes for Cpp compilation has changed.
They create a CppCompilationHandler object and use its methods as arguments in compile() function.