Backend: Module compile_golang implements a subclass GolangCompilationHandler of CompilationHandler abstract class
The existing functions in compile_golang module have become the implementation of the abstract methods.
There is no change in the code in these functions.
Of course, the names are changed in order to match the abstract methods.
The api routes for Golang compilation have not changed.
The implementation of the api routes for Golang compilation has changed.
They create a GolangCompilationHandler object and use its methods as arguments in compile() function.